moved package to pages as it is a page
